Bug 1159486 - Make profiler use the process creation time. r=shu, r=jsantell

+++ tools/profiler/GeckoProfiler.h
@@ -165,18 +165,18 @@ static inline void profiler_responsiveness(const mozilla::TimeStamp& aTime) {}
 static inline void profiler_set_frame_number(int frameNumber) {}
 
 // Get the profile encoded as a JSON string.
-static inline mozilla::UniquePtr<char[]> profiler_get_profile(float aSinceTime = 0) {
+static inline mozilla::UniquePtr<char[]> profiler_get_profile(double aSinceTime = 0) {
   return nullptr;
 }
 
 // Get the profile encoded as a JSON object.
 static inline JSObject* profiler_get_profile_jsobject(JSContext* aCx,
-                                                      float aSinceTime = 0) {
+                                                      double aSinceTime = 0) {
   return nullptr;
 }
 
 // Get the profile encoded as a JSON object.
-static inline void profiler_get_profile_jsobject_async(float aSinceTime = 0,
+static inline void profiler_get_profile_jsobject_async(double aSinceTime = 0,
                                                        mozilla::dom::Promise* = 0) {}
 
 // Get the profile and write it into a file
